pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "unbihexium"
version = "1.0.0"
description = "Production-grade Earth Observation, Geospatial, Remote Sensing, and SAR Python library"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Unbihexium OSS Foundation", email = "opensource@unbihexium.org"},
]
maintainers = [
    {name = "Unbihexium OSS Foundation", email = "opensource@unbihexium.org"},
]
keywords = [
    "earth-observation",
    "geospatial",
    "remote-sensing",
    "sar",
    "satellite-imagery",
    "gis",
    "machine-learning",
    "deep-learning",
    "object-detection",
    "segmentation",
    "change-detection",
    "flood-mapping",
    "agriculture",
    "urban-planning",
    "environmental-monitoring",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "rasterio>=1.3.0",
    "shapely>=2.0.0",
    "geopandas>=0.14.0",
    "pyproj>=3.6.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "pillow>=10.0.0",
    "scikit-learn>=1.3.0",
    "scikit-image>=0.22.0",
]

[project.optional-dependencies]
torch = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
]
gpu = [
    "unbihexium[torch]",
    "cupy-cuda12x>=12.0.0",
]
dask = [
    "dask[complete]>=2024.1.0",
]
ray = [
    "ray>=2.9.0",
]
zarr = [
    "zarr>=2.16.0",
]
netcdf = [
    "netCDF4>=1.6.0",
    "h5py>=3.10.0",
]
stac = [
    "pystac>=1.9.0",
    "pystac-client>=0.7.0",
]
parquet = [
    "pyarrow>=14.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "pymdown-extensions>=10.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "ruff>=0.1.0",
    "pyright>=1.1.340",
    "pre-commit>=3.6.0",
    "bandit>=1.7.0",
    "pip-audit>=2.6.0",
]
all = [
    "unbihexium[torch,dask,ray,zarr,netcdf,stac,parquet,docs,dev]",
]

[project.scripts]
unbihexium = "unbihexium.cli.main:main"

[project.urls]
Homepage = "https://unbihexium.org"
Documentation = "https://unbihexium.org/docs"
Repository = "https://github.com/unbihexium-oss/unbihexium"
Issues = "https://github.com/unbihexium-oss/unbihexium/issues"
Changelog = "https://github.com/unbihexium-oss/unbihexium/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/model_zoo",
    "/LICENSE",
    "/README.md",
    "/CHANGELOG.md",
]

[tool.hatch.build.targets.wheel]
packages = ["src/unbihexium"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
    "ANN",    # flake8-annotations
    "D",      # pydocstyle
    "S",      # flake8-bandit
]
ignore = [
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D107",   # Missing docstring in __init__
    "ANN101", # Missing type annotation for self
    "ANN102", # Missing type annotation for cls
    "S101",   # Use of assert detected
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["unbihexium"]

[tool.pyright]
include = ["src"]
pythonVersion = "3.10"
typeCheckingMode = "strict"
reportMissingImports = true
reportMissingTypeStubs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests that require GPU",
    "integration: marks integration tests",
]

[tool.coverage.run]
source = ["src/unbihexium"]
branch = true
parallel = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "benchmarks"]
skips = ["B101"]
